requirements.txt
flask>=2.3
scikit-learn>=1.3
numpy>=1.24
Pillow>=10.0
matplotlib>=3.7
seaborn>=0.12